resolve code review
