+++ vllm/model_executor/models/interfaces.py
@@ -1503,22 +1503,6 @@ def supports_xdrope(
     return isinstance(model, SupportsXDRoPE)
 
 
-def scatter_output_slices(
-    output: torch.Tensor,
-    indices: list[int],
-    per_item_out_tokens: list[int],
-    dest: dict[int, torch.Tensor] | list[torch.Tensor | None],
-    clone: bool = False,
-) -> None:
-    """Slice a concatenated output tensor and scatter into dest by index."""
-    offset = 0
-    for idx in indices:
-        n_tok = per_item_out_tokens[idx]
-        sliced = output[offset : offset + n_tok]
-        dest[idx] = sliced.clone() if clone else sliced
-        offset += n_tok
-
-
 @runtime_checkable
 class SupportsEncoderCudaGraph(Protocol):
     """Interface for models whose vision encoder supports CUDA graph
@@ -1626,6 +1610,8 @@ class SupportsEncoderCudaGraph(Protocol):
         encoder output prior to scattering, e.g. Step3-VL, which merges features
         according to dynamic patch counts before scattering.
         """
+        from vllm.model_executor.models.utils import scatter_output_slices
+
         scatter_output_slices(output, indices, per_item_out_tokens, dest, clone)
 
     def prepare_encoder_cudagraph_capture_inputs(
